backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
cryptography==46.0.6
dj-rest-auth==7.2.0
Django==6.0.2
django-allauth==65.15.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
inflection==0.5.1
idna==3.11
packaging==26.0
pandas==2.2.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
PyYAML==6.0.3
python-dotenv==1.2.1
requests==2.32.5
scikit-surprise==1.1.4
setuptools==82.0.0
sqlparse==0.5.5
urllib3==2.6.3
wheel==0.46.3
numpy>=1.19.5,<2
numpy<2
